Move all query stuff to api module instead of Bukkit.
