+++ quickshop-api/src/main/java/com/ghostchu/quickshop/api/shop/query/filters/MaterialFilter.java
@@ -1,4 +1,4 @@
-package com.ghostchu.quickshop.shop.query.filters;
+package com.ghostchu.quickshop.api.shop.query.filters;
 
 /*
  * QuickShop-Hikari
@@ -18,19 +18,18 @@ package com.ghostchu.quickshop.shop.query.filters;
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import com.ghostchu.quickshop.api.shop.IShopType;
 import com.ghostchu.quickshop.api.shop.Shop;
 import com.ghostchu.quickshop.api.shop.query.Filter;
-import org.jetbrains.annotations.NotNull;
+import org.bukkit.Material;
 import org.jspecify.annotations.NonNull;
 
 /**
- * TypeFilter
+ * MaterialFilter
  *
  * @author creatorfromhell
  * @since 6.3.0.0
  */
-public class TypeFilter implements Filter<IShopType> {
+public class MaterialFilter implements Filter<Material> {
 
   /**
    * Determines whether the specified object satisfies the criteria defined by the implementation in
@@ -43,8 +42,8 @@ public class TypeFilter implements Filter<IShopType> {
    * otherwise.
    */
   @Override
-  public boolean applies(final @NotNull Shop shop, @NonNull final IShopType object) {
+  public boolean applies(final @NonNull Shop shop, final @NonNull Material object) {
 
-    return shop.shopType().identifier().equalsIgnoreCase(object.identifier());
+    return shop.getItem().getType() == object;
   }
 }
